move shutdown into crate
