+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
 	"crates/httpclient",
 	"crates/pg",
 	"crates/service",
+	"crates/shutdown",
 	"crates/types",
 	"services/graphql",
 	"services/rule",
