+++ services/rule/Cargo.toml
@@ -3,13 +3,13 @@ name = "rule"
 version = "0.1.0"
 edition = "2021"
 publish = false
+rust-version.workspace = true
 
 [dependencies]
 axum = "0.6.2"
 tokio = { version = "1.33.0", default-features = false, features = [
 	"rt-multi-thread",
 	"macros",
-	"signal"
 ] }
 tracing = "0.1.37"
 tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
@@ -18,6 +18,7 @@ pg = { path = "../../crates/pg" }
 chrono = "0.4.23"
 nanoid = "0.4.0"
 service = { path = "../../crates/service" }
+shutdown = { path = "../../crates/shutdown" }
 
 [dev-dependencies]
 regex = "1.0.0"
